Split out server parts, simplify client JS layout, add listeners
This removes the separate JS client for the host user, instead opting to rely on conditionals to control whether the client should be sending messages that control the game.
It also introduces the listener paradigm to the server, such that other parts of the program listen for events and update status, rather than tightly coupling the gameplay into the server package.
The server package has been split up into separate files that should be easier to understand.