Add Selecting player tracking and player rejoining
This adds the ability to track which player is currently selecting a question. If a player gets a question correct, they get to pick the next question. It is random at first.
If a player refreshes or gets disconnected ,they can now rejoin the game by using the same passcode.
Also, the Javascript has been slightly refactored to use the event bus object to pass around the starting of a countdown, instead of the messy incrementing counters.